Refactor: extract metadata build to a class MetadataPayloadBuilder
