+++ i18n/en.pot
@@ -5,8 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2025-04-24T05:27:13.891Z\n"
-"PO-Revision-Date: 2025-04-24T05:27:13.891Z\n"
+"POT-Creation-Date: 2025-04-28T05:48:41.576Z\n"
+"PO-Revision-Date: 2025-04-28T05:48:41.576Z\n"
 
 msgid ""
 "THIS NEW RELEASE INCLUDES SHARING SETTINGS PER INSTANCES. FOR THIS VERSION "
