+++ src/domain/metadata/builders/__tests__/MetadataSyncUseCase.spec.tsx
@@ -2,7 +2,6 @@ import { anything, instance, mock, when } from "ts-mockito";
 import { RepositoryFactory } from "../../../common/factories/RepositoryFactory";
 import { Instance } from "../../../instance/entities/Instance";
 import { MetadataRepository } from "../../repositories/MetadataRepository";
-import { MetadataSyncUseCase } from "../MetadataSyncUseCase";
 import { SynchronizationBuilder } from "../../../synchronization/entities/SynchronizationBuilder";
 import { InstanceRepository } from "../../../instance/repositories/InstanceRepository";
 import { SynchronizationPayload } from "../../../synchronization/entities/SynchronizationPayload";
@@ -30,11 +29,12 @@ import {
     getDataSetMetadataByIdsResponsesWithIncludeAll,
     getDataSetTypeExpectedPayload,
 } from "./data/data-set-metadata-type";
+import { MetadataPayloadBuilder } from "../MetadataPayloadBuilder";
 
 // TODO: Notice these tests are fragile and can break easily if MetadataSyncUseCase or the metadata structure changes.
 // It is necesary a refactor of MetadataSyncUseCase and the tests to make them more robust.
 describe("MetadataSyncUseCase", () => {
-    describe("executing buildPayload method for a Category metadata type with default dependencies", () => {
+    describe("executing build method for a Category metadata type with default dependencies", () => {
         it("should return expected payload when option include objects and references of sharing settings, users and organisation units is selected", async () => {
             const includeObjectsAndReferencesOptions = {
                 includeObjectsAndReferences: true,
@@ -45,7 +45,7 @@ describe("MetadataSyncUseCase", () => {
 
             const metadataSyncUseCase = givenMetadataSyncUseCaseOfCategory(builder, includeObjectsAndReferencesOptions);
 
-            const payload: SynchronizationPayload = await metadataSyncUseCase.buildPayload();
+            const payload: SynchronizationPayload = await metadataSyncUseCase.build();
 
             const expectedPayload: SynchronizationPayload = getCategoryTypeExpectedPayload(
                 includeObjectsAndReferencesOptions
@@ -64,7 +64,7 @@ describe("MetadataSyncUseCase", () => {
 
             const metadataSyncUseCase = givenMetadataSyncUseCaseOfCategory(builder, includeObjectsAndReferencesOptions);
 
-            const payload: SynchronizationPayload = await metadataSyncUseCase.buildPayload();
+            const payload: SynchronizationPayload = await metadataSyncUseCase.build();
 
             const expectedPayload: SynchronizationPayload = getCategoryTypeExpectedPayload(
                 includeObjectsAndReferencesOptions
@@ -83,7 +83,7 @@ describe("MetadataSyncUseCase", () => {
 
             const metadataSyncUseCase = givenMetadataSyncUseCaseOfCategory(builder, includeObjectsAndReferencesOptions);
 
-            const payload: SynchronizationPayload = await metadataSyncUseCase.buildPayload();
+            const payload: SynchronizationPayload = await metadataSyncUseCase.build();
 
             const expectedPayload: SynchronizationPayload = getCategoryTypeExpectedPayload(
                 includeObjectsAndReferencesOptions
@@ -98,7 +98,7 @@ describe("MetadataSyncUseCase", () => {
                 includeObjectsAndReferences: boolean;
                 includeOnlyReferences: boolean;
             }
-        ): MetadataSyncUseCase {
+        ): MetadataPayloadBuilder {
             const { includeObjectsAndReferences } = options;
 
             const mockedInstanceRepository = mock<InstanceRepository>();
@@ -144,17 +144,17 @@ describe("MetadataSyncUseCase", () => {
             when(mockedRepositoryFactory.instanceRepository(anything())).thenReturn(instance(mockedInstanceRepository));
             when(mockedRepositoryFactory.metadataRepository(anything())).thenReturn(instance(mockedMetadataRepository));
 
-            const metadataSyncUseCase = new MetadataSyncUseCase(
+            const metadataPayloadBuilder = new MetadataPayloadBuilder(
                 builder,
                 instance(mockedRepositoryFactory),
                 dummyInstance
             );
 
-            return metadataSyncUseCase;
+            return metadataPayloadBuilder;
         }
     });
 
-    describe("executing buildPayload method for a Program metadata type with default dependencies", () => {
+    describe("executing build method for a Program metadata type with default dependencies", () => {
         it("should return expected payload when option include objects and references of sharing settings, users and organisation units is selected", async () => {
             const includeObjectsAndReferencesOptions = {
                 includeObjectsAndReferences: true,
@@ -165,7 +165,7 @@ describe("MetadataSyncUseCase", () => {
 
             const metadataSyncUseCase = givenMetadataSyncUseCaseOfProgram(builder, includeObjectsAndReferencesOptions);
 
-            const payload: SynchronizationPayload = await metadataSyncUseCase.buildPayload();
+            const payload: SynchronizationPayload = await metadataSyncUseCase.build();
 
             const expectedPayload: SynchronizationPayload = getProgramTypeExpectedPayload(
                 includeObjectsAndReferencesOptions
@@ -184,7 +184,7 @@ describe("MetadataSyncUseCase", () => {
 
             const metadataSyncUseCase = givenMetadataSyncUseCaseOfProgram(builder, includeObjectsAndReferencesOptions);
 
-            const payload: SynchronizationPayload = await metadataSyncUseCase.buildPayload();
+            const payload: SynchronizationPayload = await metadataSyncUseCase.build();
 
             const expectedPayload: SynchronizationPayload = getProgramTypeExpectedPayload(
                 includeObjectsAndReferencesOptions
@@ -203,7 +203,7 @@ describe("MetadataSyncUseCase", () => {
 
             const metadataSyncUseCase = givenMetadataSyncUseCaseOfProgram(builder, includeObjectsAndReferencesOptions);
 
-            const payload: SynchronizationPayload = await metadataSyncUseCase.buildPayload();
+            const payload: SynchronizationPayload = await metadataSyncUseCase.build();
 
             const expectedPayload: SynchronizationPayload = getProgramTypeExpectedPayload(
                 includeObjectsAndReferencesOptions
@@ -218,7 +218,7 @@ describe("MetadataSyncUseCase", () => {
                 includeObjectsAndReferences: boolean;
                 includeOnlyReferences: boolean;
             }
-        ): MetadataSyncUseCase {
+        ): MetadataPayloadBuilder {
             const { includeObjectsAndReferences } = options;
 
             const mockedInstanceRepository = mock<InstanceRepository>();
@@ -276,17 +276,17 @@ describe("MetadataSyncUseCase", () => {
             when(mockedRepositoryFactory.instanceRepository(anything())).thenReturn(instance(mockedInstanceRepository));
             when(mockedRepositoryFactory.metadataRepository(anything())).thenReturn(instance(mockedMetadataRepository));
 
-            const metadataSyncUseCase = new MetadataSyncUseCase(
+            const metadataPayloadBuilder = new MetadataPayloadBuilder(
                 builder,
                 instance(mockedRepositoryFactory),
                 dummyInstance
             );
 
-            return metadataSyncUseCase;
+            return metadataPayloadBuilder;
         }
     });
 
-    describe("executing buildPayload method for a DataSet metadata type with default dependencies", () => {
+    describe("executing build method for a DataSet metadata type with default dependencies", () => {
         it("should return expected payload when option include objects and references of sharing settings, users and organisation units is selected", async () => {
             const includeObjectsAndReferencesOptions = {
                 includeObjectsAndReferences: true,
@@ -297,7 +297,7 @@ describe("MetadataSyncUseCase", () => {
 
             const metadataSyncUseCase = givenMetadataSyncUseCaseOfDataSet(builder, includeObjectsAndReferencesOptions);
 
-            const payload: SynchronizationPayload = await metadataSyncUseCase.buildPayload();
+            const payload: SynchronizationPayload = await metadataSyncUseCase.build();
 
             const expectedPayload: SynchronizationPayload = getDataSetTypeExpectedPayload(
                 includeObjectsAndReferencesOptions
@@ -316,7 +316,7 @@ describe("MetadataSyncUseCase", () => {
 
             const metadataSyncUseCase = givenMetadataSyncUseCaseOfDataSet(builder, includeObjectsAndReferencesOptions);
 
-            const payload: SynchronizationPayload = await metadataSyncUseCase.buildPayload();
+            const payload: SynchronizationPayload = await metadataSyncUseCase.build();
 
             const expectedPayload: SynchronizationPayload = getDataSetTypeExpectedPayload(
                 includeObjectsAndReferencesOptions
@@ -335,7 +335,7 @@ describe("MetadataSyncUseCase", () => {
 
             const metadataSyncUseCase = givenMetadataSyncUseCaseOfDataSet(builder, includeObjectsAndReferencesOptions);
 
-            const payload: SynchronizationPayload = await metadataSyncUseCase.buildPayload();
+            const payload: SynchronizationPayload = await metadataSyncUseCase.build();
 
             const expectedPayload: SynchronizationPayload = getDataSetTypeExpectedPayload(
                 includeObjectsAndReferencesOptions
@@ -350,7 +350,7 @@ describe("MetadataSyncUseCase", () => {
                 includeObjectsAndReferences: boolean;
                 includeOnlyReferences: boolean;
             }
-        ): MetadataSyncUseCase {
+        ): MetadataPayloadBuilder {
             const { includeObjectsAndReferences } = options;
 
             const mockedInstanceRepository = mock<InstanceRepository>();
@@ -403,13 +403,13 @@ describe("MetadataSyncUseCase", () => {
             when(mockedRepositoryFactory.instanceRepository(anything())).thenReturn(instance(mockedInstanceRepository));
             when(mockedRepositoryFactory.metadataRepository(anything())).thenReturn(instance(mockedMetadataRepository));
 
-            const metadataSyncUseCase = new MetadataSyncUseCase(
+            const metadataPayloadBuilder = new MetadataPayloadBuilder(
                 builder,
                 instance(mockedRepositoryFactory),
                 dummyInstance
             );
 
-            return metadataSyncUseCase;
+            return metadataPayloadBuilder;
         }
     });
